remove dublicate tests
