+++ tests/Telemetry.Api.IntegrationTests/MongoDbIntegrationTests.cs
@@ -4,6 +4,7 @@ using Testcontainers.MongoDb;
 
 namespace Telemetry.Api.IntegrationTests
 {
+    [InheritsTests]
     public class MongoDbIntegrationTests : BaseIntegrationTest
     {
         private readonly MongoDbContainer _container = new MongoDbBuilder("mongo:8.2")
@@ -25,7 +26,7 @@ namespace Telemetry.Api.IntegrationTests
             await _container.StopAsync();
             await _container.DisposeAsync();
         }
-
+        
         [Test]
         public async Task GetStatus_ReturnsOk()
         {
@@ -37,60 +38,6 @@ namespace Telemetry.Api.IntegrationTests
             string content = await response.Content.ReadAsStringAsync();
             await Assert.That(content).Contains("\"status\":\"pass\"");
             await Assert.That(content).Contains("MongoDB.EntityFrameworkCore");
-        }
-
-        [Test]
-        public async Task PostScript_ReturnsOk()
-        {
-            // Arrange
-            ScriptRecordDto dto = CreateSampleScriptDto();
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsJsonAsync("/api/v2/scripts", dto);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-
-        [Test]
-        public async Task PostEvent_ReturnsOk()
-        {
-            // Arrange
-            EventRecordDto dto = CreateSampleEventDto();
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsJsonAsync("/api/v2/events", dto);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-        
-        [Test]
-        public async Task PostRawScript_ReturnsOk()
-        {
-            // Arrange
-            var content = await File.ReadAllTextAsync("assets/script.json");
-            var stringContent = new StringContent(content, Encoding.UTF8, "application/json");
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsync("/api/v2/scripts", stringContent);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-
-        [Test]
-        public async Task PostRawEvent_ReturnsOk()
-        {
-            // Arrange
-            var content = await File.ReadAllTextAsync("assets/event.json");
-            var stringContent = new StringContent(content, Encoding.UTF8, "application/json");
-            
-            // Act
-            HttpResponseMessage response = await Client.PostAsync("/api/v2/events", stringContent);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
         }
     }
 }
