+++ tests/Telemetry.Api.IntegrationTests/SqliteIntegrationTests.cs
@@ -3,6 +3,7 @@ using Telemetry.Api.Application.DTOs;
 
 namespace Telemetry.Api.IntegrationTests
 {
+    [InheritsTests]
     public class SqliteIntegrationTests : BaseIntegrationTest
     {
         private readonly string _dbPath = $"telemetry_{Guid.NewGuid()}.db";
@@ -23,7 +24,7 @@ namespace Telemetry.Api.IntegrationTests
                 File.Delete(_dbPath);
             }
         }
-
+        
         [Test]
         public async Task GetStatus_ReturnsOk()
         {
@@ -35,60 +36,6 @@ namespace Telemetry.Api.IntegrationTests
             string content = await response.Content.ReadAsStringAsync();
             await Assert.That(content).Contains("\"status\":\"pass\"");
             await Assert.That(content).Contains("Microsoft.EntityFrameworkCore.Sqlite");
-        }
-
-        [Test]
-        public async Task PostScript_ReturnsOk()
-        {
-            // Arrange
-            ScriptRecordDto dto = CreateSampleScriptDto();
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsJsonAsync("/api/v2/scripts", dto);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-
-        [Test]
-        public async Task PostEvent_ReturnsOk()
-        {
-            // Arrange
-            EventRecordDto dto = CreateSampleEventDto();
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsJsonAsync("/api/v2/events", dto);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-        
-        [Test]
-        public async Task PostRawScript_ReturnsOk()
-        {
-            // Arrange
-            var content = await File.ReadAllTextAsync("assets/script.json");
-            var stringContent = new StringContent(content, Encoding.UTF8, "application/json");
-
-            // Act
-            HttpResponseMessage response = await Client.PostAsync("/api/v2/scripts", stringContent);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
-        }
-
-        [Test]
-        public async Task PostRawEvent_ReturnsOk()
-        {
-            // Arrange
-            var content = await File.ReadAllTextAsync("assets/event.json");
-            var stringContent = new StringContent(content, Encoding.UTF8, "application/json");
-            
-            // Act
-            HttpResponseMessage response = await Client.PostAsync("/api/v2/events", stringContent);
-
-            // Assert
-            response.EnsureSuccessStatusCode();
         }
     }
 }
